% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{BulkVolume}
\alias{BulkVolume}
\title{BulkVolume}
\usage{
BulkVolume(Xf, HorizontalLength, height)
}
\arguments{
\item{Xf}{[ft]}

\item{HorizontalLength}{[ft]}

\item{height}{[ft]}
}
\value{
a volume [ft^3]
}
\description{
BulkVolume
}
