% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{CarrKobayashiBurrowsGasViscosity}
\alias{CarrKobayashiBurrowsGasViscosity}
\title{CarrKobayashiBurrowsGasViscosity}
\usage{
CarrKobayashiBurrowsGasViscosity(
  Pressure,
  Temperature,
  SGg,
  yCO2 = 0,
  yN2 = 0,
  yH2S = 0
)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{SGg}{[]}

\item{yCO2}{= 0.0}

\item{yN2}{= 0.0}

\item{yH2S}{= 0.0}
}
\value{
A numeric scalar
}
\description{
CarrKobayashiBurrowsGasViscosity
}
