% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{CentralDifferenceMethod}
\alias{CentralDifferenceMethod}
\title{CentralDifferenceMethod}
\usage{
CentralDifferenceMethod(func, x, ...)
}
\arguments{
\item{func}{func}

\item{x}{x}

\item{...}{parameters}
}
\description{
CentralDifferenceMethod
}
