% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{DindorukChristmanSolutionGasOilRatio}
\alias{DindorukChristmanSolutionGasOilRatio}
\title{DindorukChristmanSolutionGasOilRatio}
\usage{
DindorukChristmanSolutionGasOilRatio(
  Pressure,
  Temperature,
  Pbp,
  SGo_API,
  SGg,
  Multiple = 0.2
)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Pbp}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}

\item{Multiple}{=0.2}
}
\value{
A numeric scalar
}
\description{
DindorukChristmanSolutionGasOilRatio
}
