% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{GasCompressibility}
\alias{GasCompressibility}
\title{GasCompressibility}
\usage{
GasCompressibility(Pressure, Temperature, Z)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Z}{[]}
}
\value{
A numeric scalar
}
\description{
GasCompressibility
}
