% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{GasDensity}
\alias{GasDensity}
\title{GasDensity}
\usage{
GasDensity(Pressure, Temperature, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
GasDensity
}
