% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{GasDensityFromZ}
\alias{GasDensityFromZ}
\title{GasDensityFromZ}
\usage{
GasDensityFromZ(Pressure, Temperature, SGg, Z)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{SGg}{[]}

\item{Z}{[]}
}
\value{
A numeric scalar
}
\description{
GasDensityFromZ
}
