% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{GasFormationVolumeFactor}
\alias{GasFormationVolumeFactor}
\title{GasFormationVolumeFactor}
\usage{
GasFormationVolumeFactor(Pressure, Temperature, Pb, Z, SGo_API, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Pb}{[]}

\item{Z}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
GasFormationVolumeFactor
}
