% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{GlasoSaturatedOilFormationVolumeFactor}
\alias{GlasoSaturatedOilFormationVolumeFactor}
\title{GlasoSaturatedOilFormationVolumeFactor}
\usage{
GlasoSaturatedOilFormationVolumeFactor(Rs, Temperature, SGo_API, SGg)
}
\arguments{
\item{Rs}{[]}

\item{Temperature}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
GlasoSaturatedOilFormationVolumeFactor
}
