% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{KhanOilViscosity}
\alias{KhanOilViscosity}
\title{KhanOilViscosity}
\usage{
KhanOilViscosity(P, Pb, T_R, gamma_o, gamma_g, Rs)
}
\arguments{
\item{P}{[]}

\item{Pb}{[]}

\item{T_R}{[]}

\item{gamma_o}{[]}

\item{gamma_g}{[]}

\item{Rs}{[]}
}
\value{
A numeric scalar
}
\description{
KhanOilViscosity
}
