% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{KhanOilViscosity_a}
\alias{KhanOilViscosity_a}
\title{KhanOilViscosity_a}
\usage{
KhanOilViscosity_a(mu_ob, P, Pb)
}
\arguments{
\item{mu_ob}{[]}

\item{P}{[]}

\item{Pb}{[]}
}
\value{
A numeric scalar
}
\description{
KhanOilViscosity_a
}
