% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{KhanOilViscosity_b}
\alias{KhanOilViscosity_b}
\title{KhanOilViscosity_b}
\usage{
KhanOilViscosity_b(mu_ob, P, Pb)
}
\arguments{
\item{mu_ob}{[]}

\item{P}{[]}

\item{Pb}{[]}
}
\value{
A numeric scalar
}
\description{
KhanOilViscosity_b
}
