% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{KhanOilViscosity_ob}
\alias{KhanOilViscosity_ob}
\title{KhanOilViscosity_ob}
\usage{
KhanOilViscosity_ob(T_R, gamma_o, gamma_g, Rs)
}
\arguments{
\item{T_R}{[]}

\item{gamma_o}{[]}

\item{gamma_g}{[]}

\item{Rs}{[]}
}
\value{
A numeric scalar
}
\description{
KhanOilViscosity_ob
}
