% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainGasCompressibility}
\alias{McCainGasCompressibility}
\title{McCainGasCompressibility}
\usage{
McCainGasCompressibility(Pressure, Temperature, Gammag)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Gammag}{[]}
}
\value{
A numeric scalar
}
\description{
McCainGasCompressibility
}
