% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainWaterCompressibility}
\alias{McCainWaterCompressibility}
\title{McCainWaterCompressibility
Salinity Wt\% (gm/L)}
\usage{
McCainWaterCompressibility(Pressure, Temperature, Salinity = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Salinity}{[]}
}
\description{
McCainWaterCompressibility
Salinity Wt\% (gm/L)
}
