% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainWaterDensity}
\alias{McCainWaterDensity}
\title{McCainWaterDensity
Salinity Wt\% (gm/L)}
\usage{
McCainWaterDensity(Pressure, Temperature, Salinity = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Salinity}{[]}
}
\description{
McCainWaterDensity
Salinity Wt\% (gm/L)
}
