% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainWaterFormationVolumeFactor}
\alias{McCainWaterFormationVolumeFactor}
\title{McCainWaterFormationVolumeFactor
Salinity Wt\% (gm/L)}
\usage{
McCainWaterFormationVolumeFactor(Pressure, Temperature, Salinity = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Salinity}{[]}
}
\value{
A numeric scalar
}
\description{
McCainWaterFormationVolumeFactor
Salinity Wt\% (gm/L)
}
