% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainWaterSolutionGas}
\alias{McCainWaterSolutionGas}
\title{McCainWaterSolutionGas}
\usage{
McCainWaterSolutionGas(Pressure, Temperature, Salinity = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Salinity}{= 0.0 []}
}
\value{
A numeric scalar
}
\description{
McCainWaterSolutionGas
}
