% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{McCainWaterViscosity}
\alias{McCainWaterViscosity}
\title{McCainWaterViscosity
Salinity Wt\% (gm/L)}
\usage{
McCainWaterViscosity(Pressure, Temperature, Salinity = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Salinity}{[]}
}
\description{
McCainWaterViscosity
Salinity Wt\% (gm/L)
}
