% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{MolecularWeightOil}
\alias{MolecularWeightOil}
\title{MolecularWeightOil}
\usage{
MolecularWeightOil(T_R, SGo)
}
\arguments{
\item{T_R}{[]}

\item{SGo}{[]}
}
\value{
A numeric scalar
}
\description{
MolecularWeightOil
}
