% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{OilDensity}
\alias{OilDensity}
\title{OilDensity}
\usage{
OilDensity(Rs, Bo, SGo, SGg)
}
\arguments{
\item{Rs}{[]}

\item{Bo}{[]}

\item{SGo}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
OilDensity
}
