% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{OilFormationVolumeFactor}
\alias{OilFormationVolumeFactor}
\title{OilFormationVolumeFactor}
\usage{
OilFormationVolumeFactor(Pressure, Temperature, Pb, SGo_API, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Pb}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
OilFormationVolumeFactor
}
