% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{PetroskyFarshadSaturatedOilFormationVolumeFactor}
\alias{PetroskyFarshadSaturatedOilFormationVolumeFactor}
\title{PetroskyFarshadSaturatedOilFormationVolumeFactor}
\usage{
PetroskyFarshadSaturatedOilFormationVolumeFactor(Rs, Temperature, SGo_API, SGg)
}
\arguments{
\item{Rs}{[]}

\item{Temperature}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
PetroskyFarshadSaturatedOilFormationVolumeFactor
}
