% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{PetroskySaturatedOilCompressibility}
\alias{PetroskySaturatedOilCompressibility}
\title{PetroskySaturatedOilCompressibility}
\usage{
PetroskySaturatedOilCompressibility(Pressure, Temperature, Rs, SGo_API, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Rs}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
PetroskySaturatedOilCompressibility
}
