% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{ProductionIndex}
\alias{ProductionIndex}
\title{ProductionIndex}
\usage{
ProductionIndex(k, h, Muo, Bo, a, L, rwa, skin = 0, D = 0, qo = 0)
}
\arguments{
\item{k}{[]}

\item{h}{[]}

\item{Muo}{[]}

\item{Bo}{[]}

\item{a}{[]}

\item{L}{[]}

\item{rwa}{[]}

\item{skin}{= 0 []}

\item{D}{= 0 []}

\item{qo}{= 0 []}
}
\value{
A numeric scalar
}
\description{
ProductionIndex
}
