% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{RelativePermV}
\alias{RelativePermV}
\title{RelativePermV}
\usage{
RelativePermV(T_R, Rsi, Rs, SGo)
}
\arguments{
\item{T_R}{[]}

\item{Rsi}{[]}

\item{Rs}{[]}

\item{SGo}{[]}
}
\value{
A numeric scalar
}
\description{
RelativePermV
}
