% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{SolutionGasOilRatioProduction}
\alias{SolutionGasOilRatioProduction}
\title{SolutionGasOilRatioProduction}
\usage{
SolutionGasOilRatioProduction(Rs, kro, krg, muo, mug, bo, bg)
}
\arguments{
\item{Rs}{[]}

\item{kro}{[]}

\item{krg}{[]}

\item{muo}{[]}

\item{mug}{[]}

\item{bo}{[]}

\item{bg}{[]}
}
\value{
A numeric scalar
}
\description{
SolutionGasOilRatioProduction
}
