% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{SolutionWaterOilRatioProduction}
\alias{SolutionWaterOilRatioProduction}
\title{SolutionWaterOilRatioProduction}
\usage{
SolutionWaterOilRatioProduction(kro, krw, muo, muw, bo, bw)
}
\arguments{
\item{kro}{[]}

\item{krw}{[]}

\item{muo}{[]}

\item{muw}{[]}

\item{bo}{[]}

\item{bw}{[]}
}
\value{
A numeric scalar
}
\description{
SolutionWaterOilRatioProduction
}
