% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{StandingGasPseudoCritical}
\alias{StandingGasPseudoCritical}
\title{StandingGasPseudoCritical}
\usage{
StandingGasPseudoCritical(SGg, yCO2 = 0, yN2 = 0, yH2S = 0)
}
\arguments{
\item{SGg}{[]}

\item{yCO2}{= 0.0}

\item{yN2}{= 0.0}

\item{yH2S}{= 0.0}
}
\value{
A numeric scalar
}
\description{
StandingGasPseudoCritical
}
