% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{StandingGasViscosity}
\alias{StandingGasViscosity}
\title{StandingGasViscosity}
\usage{
StandingGasViscosity(Pressure, Temperature, SGg, yCO2 = 0, yN2 = 0, yH2S = 0)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{SGg}{[]}

\item{yCO2}{= 0.0}

\item{yN2}{= 0.0}

\item{yH2S}{= 0.0}
}
\value{
A numeric scalar
}
\description{
StandingGasViscosity
}
