% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{StandingOilSolutionGas}
\alias{StandingOilSolutionGas}
\title{StandingOilSolutionGas}
\usage{
StandingOilSolutionGas(SGo_API, SGg)
}
\arguments{
\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
StandingOilSolutionGas
}
