% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{StandingSaturatedOilFormationVolumeFactor}
\alias{StandingSaturatedOilFormationVolumeFactor}
\title{StandingSaturatedOilFormationVolumeFactor}
\usage{
StandingSaturatedOilFormationVolumeFactor(Rs, Temperature, SGo_API, SGg)
}
\arguments{
\item{Rs}{[]}

\item{Temperature}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
StandingSaturatedOilFormationVolumeFactor
}
