% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{TotalCompressibility}
\alias{TotalCompressibility}
\title{TotalCompressibility}
\usage{
TotalCompressibility(
  pressure,
  temperature,
  pb,
  oilapi,
  gammag,
  so,
  sg,
  sw,
  cf = default_cf,
  salinity = 0
)
}
\arguments{
\item{pressure}{[]}

\item{temperature}{[]}

\item{pb}{[]}

\item{oilapi}{[]}

\item{gammag}{[]}

\item{so}{[]}

\item{sg}{[]}

\item{sw}{[]}

\item{cf}{=default_cf}

\item{salinity}{=0}
}
\value{
A numeric scalar
}
\description{
TotalCompressibility
}
