% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{TotalFormationVolumeFactor}
\alias{TotalFormationVolumeFactor}
\title{TotalFormationVolumeFactor}
\usage{
TotalFormationVolumeFactor(Bo, Bg, Rsoi, Rso)
}
\arguments{
\item{Bo}{[]}

\item{Bg}{[]}

\item{Rsoi}{[]}

\item{Rso}{[]}
}
\value{
A numeric scalar
}
\description{
TotalFormationVolumeFactor
}
