% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{VasquezBeggsOilSolutionGas}
\alias{VasquezBeggsOilSolutionGas}
\title{VasquezBeggsOilSolutionGas}
\usage{
VasquezBeggsOilSolutionGas(Pressure, Temperature, SGo_API, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
VasquezBeggsOilSolutionGas
}
