% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{VelardeSolutionGasOilRatio}
\alias{VelardeSolutionGasOilRatio}
\title{VelardeSolutionGasOilRatio(Rs)}
\usage{
VelardeSolutionGasOilRatio(Pressure, Temperature, Pbp, SGo_API, SGg)
}
\arguments{
\item{Pressure}{[]}

\item{Temperature}{[]}

\item{Pbp}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
VelardeSolutionGasOilRatio(Rs)
}
