% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{VelardeSolutionGasOilRatioBubblePoint}
\alias{VelardeSolutionGasOilRatioBubblePoint}
\title{VelardeSolutionGasOilRatioBubblePoint(Rsb)}
\usage{
VelardeSolutionGasOilRatioBubblePoint(Temperature, Pbp, SGo, SGg)
}
\arguments{
\item{Temperature}{[]}

\item{Pbp}{[]}

\item{SGo}{[]}

\item{SGg}{[]}
}
\value{
A numeric scalar
}
\description{
VelardeSolutionGasOilRatioBubblePoint(Rsb)
}
