% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{VolatileOilGasRatio}
\alias{VolatileOilGasRatio}
\title{VolatileOilGasRatio by El-Banbi}
\usage{
VolatileOilGasRatio(pressure, temperature, Pb, CGR, oilAPI, Gammag)
}
\arguments{
\item{pressure}{[]}

\item{temperature}{[]}

\item{Pb}{[]}

\item{CGR}{[]}

\item{oilAPI}{[]}

\item{Gammag}{[]}
}
\value{
A numeric scalar
}
\description{
VolatileOilGasRatio by El-Banbi
}
