% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{WolfcampBoi}
\alias{WolfcampBoi}
\title{WolfcampBoi(Boi)}
\usage{
WolfcampBoi(Bobp, Cobp, Pi, Pbp)
}
\arguments{
\item{Bobp}{[]}

\item{Cobp}{[]}

\item{Pi}{[]}

\item{Pbp}{[]}
}
\value{
A numeric scalar
}
\description{
WolfcampBoi(Boi)
}
