% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrocarbons.R
\name{WolfcampFVF}
\alias{WolfcampFVF}
\title{WolfcampFVF(Bobp)}
\usage{
WolfcampFVF(T_F, SGo_API, SGg, Rs)
}
\arguments{
\item{T_F}{[]}

\item{SGo_API}{[]}

\item{SGg}{[]}

\item{Rs}{[]}
}
\value{
A numeric scalar
}
\description{
WolfcampFVF(Bobp)
}
